Replace dagProba inner map with TailProba struct.
Originally, dagProba was a map[int]map[int]float64, where
the outer map tracks the index of each textRune, and the
inner map stores tail index of the slices of textRune that
are found in the prefix dictionary, and the piece's probability.
This inner map caused inconsistent results when all of their
probabilities are the same (typically when they're all equal
to minFloat). The inconsistency happens in the for loop in
maxIndexProba() because it's iterating through an UNORDERED
map. Another problem with this method is that the bestIndex
variable starts at -1, and this isn't checked before returning,
causing an infinite loop in findBestPath() method (because
i is set to a negative number at the end of each loop).