Fix infinite loop.
Found an infinite loop problem caused the the sentence
"这一刹那的撙近" from 围城.txt. The problem is manyfold.
First, the character "撙" is found in the prefix dictionary
with 0 count. This property caused the first for-loop in
buildDAG() method to do nothing for this character; in
effect leaving an empty tail slice.
Second, when findBestPath() calls maxIndexProba(), the
latter method encounters the empty slice and returns a
negative index value to the former method. findBestPath()
ends up running in an infinite loop because the i variable
is reset to -1.
Both problems are now fixed with checks for negative values.